keccak hash moved to stark transcript
